Change servlets to return Entity Objects instead of one-item ArrayLists
